+++ funcion-service/src/main/java/proyecto/funcion_service/repository/FuncionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import proyecto.funcion_service.model.Funcion;
 
 @Repository
-public interface FuncionRepository extends JpaRepository<Funcion, Long> {
+public interface FuncionRepository extends JpaRepository<Funcion, Integer> {
     // Aquí puedes agregar métodos personalizados después,
     // como buscar funciones por ID de película.
 }
